Solve: Matrix.createDeterminant(): variable D has wrong type
int D will work fine as long as the matrix content is actually ints,
but not with floats

+++ src/pallav/Matrix/Matrix.java
@@ -484,7 +484,7 @@ public class Matrix {
 	 */
 	static float createDeterminant(Matrix mat, int n) {
 		float[][] A = mat.array;
-		int D = 0;
+		float D = 0;
 		if (n == 1) {
 			return A[0][0];
 		}
